Restructured calendar module and CSS:
The calendar thumb list is now comprised of a ul/li set rather than divs.
The data display for the full event or the login/advertising is now placed in the UL directly. This enhances the responsive nature by allowing the content to automatically flow correctly on browser resize (mainly, for mobile, I think this enhances the UX).

I propose the future functionality of JS calls adding new <li> elements into the stack as the user cycles through the thumb versions, adding them just after each thumb that they pertain to. This will create a drawer-style effect that will likely be much more simple to navigate. Previously the functionality was to have the event displayed in full at the top in mobile, but that would be annoying. This data structure change has no effect on the visual layout in desktop mode except one thing:

To make this possible, the added <li> has to be absolutely positioned, which means we can't have dynamic heights on the content area anymore- the only time it would be fine is if the list of thumbs exceeded the content area's height. Since we can't rely on that, I added a min-height of 650 to the list, and forced the height of the display div to 650. I think this will be hard to notice, and we can truncate the description text for events with wordier descriptions and just force the user to click through to the event page to read on.

Too late. Going to bed.