Added new module: Tanguer_Alert.js
This is the front-end control that handles Alert functionality so the user doesn't have to refresh the page. Lacking work to actually send the "dismiss" alert via ajax, but this is fairly minor and will be done in the future.

Major GUI updates. Moved the location picker out of the header and to the right of the date selection. Site feels more balanced. Mobile CSS has to be modified to force it to wrap and center, but it is almost ready.

Did a whole bunch of code cleanup and commenting in the JS files for futureproofing.

Modified the sprite png.

Reverted logo functionality in the header. I did not  like the logo switching that I had added. Now, the logo just shrinks when you go to mobile mode by almost half. Still looks good.